Add test for score
